modification structure

+++ orderservice/src/main/java/com/idruide/backend/orderservice/service/order/OrderServiceImpl.java
@@ -4,21 +4,22 @@ package com.idruide.backend.orderservice.service.order;
 import com.idruide.backend.orderservice.dto.OrderDto;
 import com.idruide.backend.orderservice.dto.OrderProductDto;
 import com.idruide.backend.orderservice.entities.Order;
-import com.idruide.backend.orderservice.entities.OrderProduct;
 import com.idruide.backend.orderservice.entities.OrderProductPK;
 import com.idruide.backend.orderservice.entities.Product;
 import com.idruide.backend.orderservice.exception.OrderNotFoundException;
-import com.idruide.backend.orderservice.exception.ProductNotFoundException;
 import com.idruide.backend.orderservice.mapper.OrderMapper;
 import com.idruide.backend.orderservice.repository.OrderProductRepository;
 import com.idruide.backend.orderservice.repository.OrderRepository;
 import com.idruide.backend.orderservice.repository.ProductRepository;
+import com.idruide.backend.orderservice.utils.OrderDateUtils;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.rmi.server.UID;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -55,8 +56,6 @@ public class OrderServiceImpl implements OrderService {
     }
 
 
-
-
     @Override
     @Transactional
     public List<OrderDto> getAllOrders() {
@@ -71,9 +70,9 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public OrderDto getOrderByNumber(String numberOrder) {
-        return  this.orderRepository.findByOrderNumber(numberOrder)
+        return this.orderRepository.findByOrderNumber(numberOrder)
                 .stream()
-                .map(order->this.orderMapper.toOrderDto(order) )
+                .map(order -> this.orderMapper.toOrderDto(order))
                 .findFirst()
                 .orElse(null);
     }
@@ -81,40 +80,38 @@ public class OrderServiceImpl implements OrderService {
     @Override
     @Transactional
     public OrderDto saveOrder(OrderDto orderDto) {
-
         return Optional.ofNullable(orderDto)
                 .filter(this::validateOrder)
-                .map(orderDtoT ->this.orderMapper.toOrder(orderDtoT))
-                .map(order ->  {
-                    String orderNumber=String.valueOf(new UID());
+                .map(orderTemp -> this.orderMapper.toOrder(orderTemp))
+                .map(order -> {
+                    String orderNumber = String.valueOf(new UID());
                     order.setOrderNumber(orderNumber);
+                    order.setCreatedAt(OrderDateUtils.SystemDateNow());
+                    order.setDeliverDate(OrderDateUtils.estimatedDeliveryDate());
                     order.getOrderProducts().forEach(orderProduct -> {
-                        Product prod=this.productRepository.findByCodeProduct(orderProduct.getProductCode());
+                        Product prod = this.productRepository.findByCodeProduct(orderProduct.getProductCode());
                         orderProduct.setProduct(prod);
-                        //orderProduct.setOrder(order);});
                         orderProduct.setOrderNumber(orderNumber);
                         orderProduct.setId(OrderProductPK.builder().orderNumber(orderNumber).product(prod).build());
-
+                        orderProduct.setOrderId(order);
                     });
-
                     return order;
                 })
-                .map( order -> this.orderRepository.save(order))
+                .map(order -> this.orderRepository.save(order))
                 .map(order -> this.orderMapper.toOrderDto(order))
                 .orElse(null);
-
     }
 
     boolean validateOrder(OrderDto orderDto) {
-        List <OrderProductDto> validated = Optional.ofNullable(orderDto.getOrderProducts())
+        List<OrderProductDto> validated = Optional.ofNullable(orderDto.getOrderProducts())
                 .map(List::stream)
                 .orElseGet(Stream::empty)
                 .map(orderProduct -> {
                     Integer avalaibleQuantity = Optional.ofNullable(this.productRepository.findByCodeProduct(orderProduct.getCodeProduct()))
                             .orElse(Product.builder().quantity(0).build()).getQuantity();
-                    if(avalaibleQuantity > orderProduct.getQuantity()){
+                    if (avalaibleQuantity > orderProduct.getQuantity()) {
                         return orderProduct;
-                    }else {
+                    } else {
                         return null;
                     }
                 })
@@ -124,37 +121,57 @@ public class OrderServiceImpl implements OrderService {
 
     }
 
-    @Override
-    @Transactional
-    public OrderDto updateOrder(Integer orderId, Integer productId) {
-        Order order = this.orderRepository.findById(orderId).orElseThrow(() -> new OrderNotFoundException("Order not found", orderId));
-        OrderProduct product = this.orderProductRepository.findById(productId).orElseThrow(() -> new ProductNotFoundException("Products not found", productId));
-        //order.addProducts(product);
-        return this.orderMapper.toOrderDto(this.orderRepository.save(order));
+    public OrderDto deleteOrder(OrderDto orderDto) {
+        return Optional.ofNullable(orderDto.getOrderNumber())
+                .filter(StringUtils::isNotBlank)
+                .map(ordNumber -> this.orderRepository.findByOrderNumber(ordNumber))
+                .map(orders -> {
+                    if (orders.isEmpty()) {
+                        return null;
+                    } else return orders;
+                })
+                .map(List::stream)
+                .get()
+                .map(order -> {
+                    if (order != null) this.orderRepository.delete(order);
+                    return this.orderMapper.toOrderDto(order);
+                })
+                .findFirst().orElseThrow(() -> new OrderNotFoundException("order Number not found. ", -1));
     }
 
     @Override
     @Transactional
-    public OrderDto saveOrderProduct(OrderDto orderDto) {
-        Order order = this.orderMapper.toOrderProduct(orderDto);
-        this.orderRepository.save(order);
-        return orderDto;
-    }
-
-    @Override
-    @Transactional
-    public void deleteOrder(OrderDto orderDto) {
-        //get orderNumber
-        String orderNumber = orderDto.getOrderNumber();
-        //if not
-        if(orderNumber==null){
-            log.warn("order: not found. ");
-            new OrderNotFoundException("Order: not found. ",-1);
-        }
-        //else if get order by number
-        OrderDto orderDtoRec = getOrderByNumber(orderNumber);
-        Order order = this.orderMapper.toOrder(orderDtoRec);
-        this.orderRepository.delete(order);
-    }
-
+    public OrderDto updateOrder(OrderDto orderDto) {
+        return  Optional.ofNullable(orderDto.getOrderNumber())
+                .filter(StringUtils::isNotBlank)
+                .map(orderNumber -> orderRepository.findByOrderNumber(orderNumber))
+                .map(orders -> {
+                    if (orders.isEmpty()) {
+                        Order ord = this.orderMapper.toOrder(orderDto);
+                        return new ArrayList<Order>() {{
+                            add(ord);
+                        }};
+                    } else return orders;
+                })
+                .map(List::stream)
+                .get()
+                .map(order -> {
+                    order.setCreatedAt(OrderDateUtils.SystemDateNow());
+                    order.setDeliverDate(OrderDateUtils.estimatedDeliveryDate());
+                    order.getOrderProducts().forEach(orderProduct -> {
+                        Product prod = this.productRepository.findByCodeProduct(orderProduct.getProductCode());
+                        orderProduct.setProduct(prod);
+                        orderProduct.setOrderNumber(orderDto.getOrderNumber());
+                        orderProduct.setId(OrderProductPK.builder().orderNumber(orderDto.getOrderNumber()).product(prod).build());
+                        orderProduct.setOrderId(order);
+                    });
+                    return order;
+                })
+                .map(order -> this.orderRepository.save(order))
+                .map(order -> this.orderMapper.toOrderDto(order))
+                .findFirst().orElseThrow(() -> new OrderNotFoundException("order not found. ", -1));
+       }
 }
+
+
+
